requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.5.0
groq==0.8.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.1
idna==3.7
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-core==0.2.1
langchain-groq==0.1.4
langchain-mistralai==0.1.7
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langchainhub==0.1.16
langsmith==0.1.63
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
openai==1.30.3
orjson==3.10.3
packaging==23.2
pedalboard==0.9.6
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
tenacity==8.3.0
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
typer==0.12.3
types-requests==2.32.0.20240523
typing_extensions==4.12.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
